Send multi-line mutants to the frontend
For the frontend changes: issue with mutationResult.mutatedLines (Highlighter.tsx:176) being undefined. This is because in run.ts:90 mutant has no attribute mutatedLines, we have addedLines and removedLines now. Change to addedLines (before, regex was just looking for + so I figured this was the right call). Still erroring because it's not in the db and I was instructed not to change the db